refractored bismark module
